feat: add authentication and role based authorization! created nextjs middleware to ran through it's middleware function on the routes specified in config matcher. if it matches.. it will ru the middleware function which run the logic
in this example, if the pathname starts with '/users/', and the request's token role is not ADMIN, then the page will be rewritten with '/access-denied' page, but with the original url still showing.

and if the samething also applies to the page that starts with '/articles', where the request's token role mustbe either ADMIN or USER to view it.

this is just an example tho.. I'll finnish the middleware after i finnished my add articles page.
cuz the whole reason i did this auth bit is for me to get access to the current state of who is logged in!
cuz like i need the info about who the frick is creating the articles lol.
but not i can! by accessing the session's id that i assigned on the CredentialsProvider callback!

noiceee